requirements.txt
djangorestframework
django==5.2
pyyaml
django-filter
django-oauth-toolkit
